Ship multi-app polish: customer orders, owner onboarding, delivery boy, and billing UX.
Improve add-customer subscription flow, billing toolbar layout, delivery boy screens, and related API/notifications so owner, customer, and delivery apps stay aligned for device testing.

+++ lib/features/customer/data/repositories/customer_order_repository.dart
@@ -65,7 +65,7 @@ class CustomerOrderRepository {
   Future<void> updateQty(
     String date,
     int subscriptionLineId,
-    int qty,
+    double qty,
   ) async {
     final response = await _dio.put<Map<String, dynamic>>(
       '/customer/v1/orders/$date/qty',
